# Performance improvement: avoid copying govaluate parameters unnecessarily