@jszwedko 2019-06-27
CodeClimate seems to be complaining about the number of `return` values in `enforceParameters.Get` but I think what I have is idiomatic Go. Happy to refactor if desired though.